# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 🚀 Edu Management API
  description: Education Management System API collection
  version: 1.0.0
servers:
  - url: http://{{base_url}}{{api_version}}
  - url: http://{{base_url}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...
paths:
...
  /users/register/student:
    post:
      tags:
        - 📝 Register
      summary: Register Student
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"email\": \"student.new@school.com\",\n  \"password\":
                \"Password@123\",\n  \"admission_no\": \"ADM2024010\",\n 
                \"first_name\": \"Rahul\",\n  \"last_name\": \"Verma\",\n 
                \"dob\": \"2010-05-15T00:00:00Z\",\n  \"gender\": \"MALE\",\n 
                \"phone\": \"9123456799\",\n  \"blood_group\": \"B+\",\n 
                \"admission_date\": \"2024-06-01T00:00:00Z\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /students/:
    get:
      tags:
        - 👨‍🎓 Students
      summary: List Students
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: limit
          in: query
          schema:
            type: integer
          example: '10'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /students/{id}:
    get:
      tags:
        - 👨‍🎓 Students
      summary: Get Student By ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - 👨‍🎓 Students
      summary: Update Student
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"first_name\": \"Arjun\",\n  \"last_name\":
                \"Sharma\",\n  \"phone\": \"9123456789\",\n  \"address\": \"123
                Main Street, Chennai\",\n  \"blood_group\": \"O+\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 👨‍🎓 Students
      summary: Delete Student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /enrollments/student/{student_id}:
    get:
      tags:
        - 📑 Enrollment
      summary: Get Enrollments By Student
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: limit
          in: query
          schema:
            type: integer
          example: '10'
        - name: student_id
          in: path
          schema:
            type: string
          required: true
          example: STUDENT_ID_HERE
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /standards/:
    post:
      tags:
        - 📏 Standard
      summary: Create Standard
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"name\": \"Grade 1\",\n  \"department_id\":
                \"DEPARTMENT_ID_HERE\",\n  \"order_index\": 1,\n 
                \"description\": \"First grade\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - 📏 Standard
      summary: List All Standards
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /standards/{id}:
    get:
      tags:
        - 📏 Standard
      summary: Get Standard By ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - 📏 Standard
      summary: Update Standard
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"name\": \"Grade 1\",\n  \"order_index\": 1,\n 
                \"description\": \"First standard\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 📏 Standard
      summary: Delete Standard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /standards/department/{department_id}:
    get:
      tags:
        - 📏 Standard
      summary: List Standards By Department
      security:
        - bearerAuth: []
      parameters:
        - name: department_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /standards/{id}/subjects:
    get:
      tags:
        - 📏 Standard
      summary: Get Subjects For Standard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - 📏 Standard
      summary: Assign Subject To Standard
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"subject_id\": \"SUBJECT_ID_HERE\",\n  \"is_core\":
                true\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /standards/{id}/subjects/{subject_id}:
    delete:
      tags:
        - 📏 Standard
      summary: Remove Subject From Standard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
        - name: subject_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /class-sections/standard/{standard_id}:
    get:
      tags:
        - 🏫 Class Section
      summary: List Class Sections By Standard
      security:
        - bearerAuth: []
      parameters:
        - name: X-Academic-Year-ID
          in: header
          schema:
            type: string
          example: '{{academic_year_id}}'
        - name: standard_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /subjects/:
    post:
      tags:
        - 📚 Subject
      summary: Create Subject
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"code\": \"MATH101\",\n  \"name\": \"Mathematics\",\n 
                \"description\": \"Core mathematics\",\n  \"is_elective\":
                false\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - 📚 Subject
      summary: List Subjects — Search
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          example: math
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /subjects/{id}:
    get:
      tags:
        - 📚 Subject
      summary: Get Subject By ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - 📚 Subject
      summary: Update Subject
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"name\": \"Advanced Mathematics\",\n  \"description\":
                \"Advanced core mathematics\",\n  \"is_elective\": false\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 📚 Subject
      summary: Delete Subject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /exams-result/student/{enrollment_id}:
    get:
      tags:
        - 📊 Exam Result
      summary: List Results by Student Enrollment
      parameters:
        - name: enrollment_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /fees/structure/standard/{standard_id}:
    get:
      tags:
        - 💰 Fee > Structure
      summary: Get Fee Structures By Standard
      description: Returns all fee components for a standard in the current academic year.
      security:
        - bearerAuth: []
      parameters:
        - name: X-Academic-Year-ID
          in: header
          schema:
            type: string
          example: '{{academic_year_id}}'
        - name: standard_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /reports/attendance/student:
    get:
      tags:
        - 📈 Reports
      summary: Student Attendance — Date Range
      security:
        - bearerAuth: []
      parameters:
        - name: student_enrollment_id
          in: query
          schema:
            type: string
          example: ENROLLMENT_ID_HERE
        - name: from_date
          in: query
          schema:
            type: string
          example: '2024-06-01'
        - name: to_date
          in: query
          schema:
            type: string
          example: '2024-09-30'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...